+++ fawlty/exception.py
