+++ fawlty/resources/asset.py
@@ -6,7 +6,7 @@ A module to represent a Sensu asset resource
 from typing import Optional, Dict, List
 
 # Our imports
-from sensu.resources.base import ResourceBase
+from sensu.resources.base import ResourceBase, MetadataWithNamespace
 from sensu.client import SensuClient
 
 # 3rd party imports
@@ -27,15 +27,10 @@ def get_url(namespace: str, name: str = None) -> str:
     return url
 
 
-class AssetMetadata(BaseModel):
+class AssetMetadata(MetadataWithNamespace):
     """
     A class to represent the data structure of an asset metadata
     """
-    name: str
-    namespace: str
-    created_by: Optional[str] = None
-    labels: Optional[Dict[str, str]] = {}
-    annotations: Optional[Dict[str, str]] = {}
 
 
 class Asset(ResourceBase):
