+++ fawlty/resources/clusterrolebinding.py
@@ -5,7 +5,7 @@ A module to represent a Sensu clusterrolebinding resource
 from typing import Optional, List, Dict, Literal
 
 # Our imports
-from sensu.resources.base import ResourceBase
+from sensu.resources.base import ResourceBase, MetadataWithoutNamespace
 from sensu.client import SensuClient
 
 # 3rd party imports
@@ -26,14 +26,10 @@ def get_url(name: str = None) -> str:
     return url
 
 
-class ClusterRoleBindingMetadata(BaseModel):
+class ClusterRoleBindingMetadata(MetadataWithoutNamespace):
     """
     A class to represent the data structure of a clusterrole-binding metadata
     """
-    name: str
-    created_by: Optional[str] = None
-    labels: Optional[dict[str, str]] = {}
-    annotations: Optional[dict[str, str]] = {}
 
 
 class ClusterRoleBindingSubject(BaseModel):
