+++ fawlty/resources/handler.py
@@ -6,7 +6,7 @@ A module for Sensu handler resources.
 from typing import Optional, List, Dict, Literal
 
 # Our imports
-from sensu.resources.base import ResourceBase
+from sensu.resources.base import ResourceBase, MetadataWithNamespace
 from sensu.client import SensuClient
 
 # 3rd party imports
@@ -27,15 +27,10 @@ def get_url(namespace: str, name: str = None) -> str:
     return url
 
 
-class HandlerMetadata(BaseModel):
+class HandlerMetadata(MetadataWithNamespace):
     """
     A class to represent the data structure of a handler metadata
     """
-    name: str
-    namespace: str
-    created_by: Optional[str] = None
-    labels: Optional[dict[str, str]] = {}
-    annotations: Optional[dict[str, str]] = {}
 
 
 class HandlerSocket(BaseModel):
