+++ fawlty/resources/namespace.py
